test: combine test suites in different test files
